# MyntraProject/Myntra/src/main/java/com/ms/myntra/entity/MyntraPayload.java
package com.ms.myntra.entity;

public class MyntraPayload {
	
	private String alarmType;
	
	private String alarm;
	
	private String clearUserId;
	
	private String Notes;
	
	private String Id;


	public String getAlarmType() {
		return alarmType; 
	}

	public void setAlarmType(String alarmType) {
		this.alarmType = alarmType;
	}

	public String getAlarm() {
		return alarm;
	}

	public void setAlarm(String alarm) {
		this.alarm = alarm;
	}

	public String getClearUserId() {
		return clearUserId;
	}

	public void setClearUserId(String clearUserId) {
		this.clearUserId = clearUserId;
	}

	public String getNotes() {
		return Notes;
	}

	public void setNotes(String notes) {
		Notes = notes;
	}

	public String getId() {
		return Id;
	}

	public void setId(String id) {
		Id = id;
	}

		

}