# MyntraProject/swagger .yaml
swagger: '2.0'
info:
  description: Validating Users through user_name and password!
  version: '1.0'
  title: Myntra Page API!!
host: 'localhost:9086'
basePath: /
tags:
  - name: code-controller
    description: Code Controller
paths:
  /myntra-page/order-creation:
    post:
      tags:
        - code-controller
      summary: myntracodeservice
      operationId: myntracodeserviceUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: myntrareq
          description: myntrareq
          required: true
          schema:
            $ref: '#/definitions/MyntraPayload'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /myntra-page/order-deletion:
    delete:
      tags:
        - code-controller
      summary: myntradeleteservice
      operationId: myntradeleteserviceUsingDELETE
      produces:
        - application/json
      parameters:
        - in: body
          name: myntrareq
          description: myntrareq
          required: true
          schema:
            $ref: '#/definitions/MyntraPayload'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  '/myntra-page/order-status/{id}':
    get:
      tags:
        - code-controller
      summary: myntragetservice
      operationId: myntragetserviceUsingGET
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /myntra-page/order-updation:
    patch:
      tags:
        - code-controller
      summary: myntrapatchservice
      operationId: myntrapatchserviceUsingPATCH
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: myntrareq
          description: myntrareq
          required: true
          schema:
            $ref: '#/definitions/MyntraPayload'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
definitions:
  MyntraPayload:
    type: object
    properties:
      alarm:
        type: string
      alarmType:
        type: string
      clearUserId:
        type: string
      id:
        type: string
      notes:
        type: string
    title: MyntraPayload